services/manager/requirements.txt
python-docx==1.1.2
pysmb==1.2.9.1